server: rocket + lazy static
The choice for the server library become rocket. It requires nighly
build, but since the aim of the project is the client application
I'm fine with the unsuppoerted features.

The state of the server will be a lazy static HashMap, which can be used
as a simple lockable module variable.