server: error handling
Replace the naive implementation with proper error hanling.
The unwrap calls can be replaced with "?" (try! macro earlier)
so the return type became Result<..., ErrorType> which caused
the map_err calls to convert the errors to the same type.

This error propagation bubbles up without losing the cause of the error,
and the sugar syntax of "?" makes it elegant (early return without extra
indentaion).