client: use reqwest as http library
The first step of the client is to fetch data from the server.

It is not so easy to find a stable http request library whithout aysnc
mechism. Most of the simple libraries don't live long. The more stable
libraries are async (tokio) heavy. Hyper seems a de facto choice for
http and reqwest is a simple wrapper over hyper which fades the future
handling from the user.

At a final step maybe the hyper library should be used directly.

However the json parsing went easy after a little reading. Since the
serde_json::Value can be used as parsing type, and it gives the required
flexibility for ruther manipulation.