client: command parsing
Command parsing had to be introduced before new user commands can be
implemented. split_whitespace and a match statement do the job fair
enough. It's not a sophisticated one, but our usecase is quite simple.
The rust infrastructure encourages simple solution.

I had a look at nom library as a lexer, but the examples were more
complicated to achive a simple goal like string matching. It may worst
to try it later.