client: command enum
The Command trait was a too early abstraction because second command
could not be added into parse function (turbofish does not support
impl), and the compile requires the same type from the match results.

Fortunately enum support our case and list command could be added.

The new features, like impl in type notation, are not supported in all
cases. We may say that the language is not coherent, but we since this
is a nightly version of the language we could not say a bad word and the
simpler solution (enum) does the job.